feat(): ✨ 异常捕获增强插件

+++ cq-core-boot-starter/src/main/java/com/dasoops/dasserver/cq/conf/properties/CqProperties.java
@@ -49,11 +49,11 @@ public class CqProperties {
     /**
      * 开发群id
      */
-    private int devGroupId;
+    private Long devGroupId;
 
     /**
      * 主开发者id
      */
-    private int devUserId;
+    private Long devUserId;
 
 }
